+++ Observations/urls.py
@@ -7,7 +7,10 @@ from Observations.views import (
     fiche_observation_view,
     fiche_test_observation_view, ajouter_observation, ajouter_remarque,
     saisie_observation, traiter_saisie_observation,
-    select_directory, process_images, check_progress, transcription_results
+)
+from Observations.views.view_transcription import (
+    select_directory, process_images, check_progress,
+    transcription_results, start_transcription_view
 )
 from Observations.views import saisie_observation_view
 
@@ -33,6 +36,7 @@ urlpatterns = [
     path('transcription/traiter-images/', process_images, name='process_images'),
     path('transcription/verifier-progression/', check_progress, name='check_progress'),
     path('transcription/resultats/', transcription_results, name='transcription_results'),
+    path('transcription/demarrer/', start_transcription_view, name='start_transcription'),
 
     # Route pour l'historique des modifications
     path('observations/historique/<int:fiche_id>/', saisie_observation_view.historique_modifications, name='historique_modifications'),